requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.3.0
aiofiles==23.2.1
gunicorn==21.2.0

# Excel processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP requests
requests==2.31.0
